Refactor issue summary to make logic reusable
